Moving Melody, Note and Instrument classes to core package
