+++ Visuals/Eclipse/workspace/CodeAndChordsEclipseProject/src/core/Melody.java
@@ -1,4 +1,4 @@
-package functionSketch_01;
+package core;
 
 import java.util.*;
